requirements.txt
flask
rembg[cpu]
opencv-python-headless
pillow
boto3
requests
cloudinary
python-dotenv
gunicorn
